+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './types';
+export * from './graphQL';
+export * from './promise';
+