+++ src/widgets/averageNumber/settings.ts
@@ -1,5 +1,4 @@
-import {IWidgetSettings, makeSettings} from "../../widgetSettings";
-import {makeColor} from "../../widgetSettings/controls";
+import {IWidgetSettings, makeSettings, makeColor} from "../../widgetSettings";
 import settingsPresets from "../../widgetSettings/settingsPresets";
 
 export const settings: IWidgetSettings = makeSettings({