+++ src/widgets/inventiveTable/index.ts
@@ -1,0 +1,2 @@
+export * from "./inventiveTable";
+export * from "./settings";