+++ src/widgets/pie/settings.ts
@@ -1,7 +1,14 @@
-import {IWidgetSettings, makeSettings} from "../../widgetSettings";
+import {
+    IWidgetSettings,
+    makeSettings,
+    makeArray,
+    makeColor,
+    makeList,
+    makeNumber,
+    makeSettingsGroup
+} from "../../widgetSettings";
 import settingsPresets from "../../widgetSettings/settingsPresets";
-import {makeArray, makeColor, makeList, makeNumber, makeSettingsGroup, makeString} from "../../widgetSettings/controls";
-import {PieLabelAlign, PieLabelAlignValues} from "../../types/types";
+import {PieLabelAlign, PieLabelAlignValues} from "../../types";
 
 export const settings: IWidgetSettings = makeSettings({
     settings: [