+++ src/widgets/productionPlan/settings.ts
@@ -1,5 +1,10 @@
-import {IWidgetSettings, makeSettings} from "../../widgetSettings";
-import {makeBoolean, makeColor, makeList, makeNumber, makeSettingsGroup, makeString} from "../../widgetSettings/controls";
+import {
+    IWidgetSettings,
+    makeSettings,
+    makeColor,
+    makeSettingsGroup,
+    makeString
+} from "../../widgetSettings";
 import settingsPresets from "../../widgetSettings/settingsPresets";
 
 export const settings: IWidgetSettings = makeSettings({